Sound Master: Merge the structs into the main .c file, disable logging and remove excess logging
